// src/interfaces/wms/transaction/inbound/packingDetails_wms.interface.ts
export interface IPackingDetailsWmsView {
  company_code: string;
  po_no: string;
  prin_code: string;
  job_no: string;
  packdet_no: number;
  prod_code: string;
  prod_name: string;
  qty_puom: number;
  p_uom: string;
  qty_luom: number;
  l_uom: string;
  quantity: number;
  container_no: string;
  clearance: string;
  allocated: string;
  updated_at?: string;
  updated_by: string;
  created_at?: string;
  created_by: string;
  MFG_DATE: Date;
  EXP_DATE: Date;
  qty_string?: string;
      uppp?: string;
}
export interface IPackingDetails {
  company_code: string;
  prin_code: string;
  job_no: string;
  packdet_no: number;
  prod_code: string;
  qty_puom: number;
  p_uom: string;
  qty_luom: number;
  l_uom: string;
  quantity: number;
  lot_no: string;
  po_no: string;
  bl_no: string;
  mfg_date: Date;
  exp_date: Date;
  vessel_name: string;
  voyage_no: string;
  container_no: string;
  invoice_value: number;
  curr_code: string;
  ex_rate: number;
  doc_ref: string;
  manu_code: string;
  from_site: string;
  to_site: string;
  from_aisle: string;
  to_aisle: string;
  from_column: number;
  to_column: number;
  from_height: number;
  to_height: number;
  selected: string;
  receipt_type: string;
  allocated: string;
  pallet_id: string;
  pallet_serial_no: string;
  mixed_putaway: string;
  user_id: string;
  user_dt: Date;
  location_from: string;
  location_to: string;
  uppp: number;
  confirmed: string;
  unit_price: number;
  apportionate_value: number;
  multi_series: string;
  clearance: string;
  cleared_date: Date;
  cleared_user: string;
  reject_reason: string;
  container_size: number;
  cust_code: string;
  moc1: string;
  moc2: string;
  simulate_flag: string;
  length: number;
  breadth: number;
  height: number;
  gross_weight: number;
  new_product: string;
  prod_name: string;
  volume: number;
  new_weight: number;
  origin_country: string;
  shelf_life_days: number;
  shelf_life_date: Date;
  pda_qty1: number;
  pda_qty2: number;
  pda_quantity: number;
  qty1_arrived: number;
  qty2_arrived: number;
  quantity_arrived: number;
  prod_attrib_code: string;
  prod_grade1: string;
  prod_grade2: string;
  tx_identity_number: string;
  net_weight: number;
  supp_code: string;
  assigned_putaway_user: string;
  assigned_tally_user: string;
  prv_location_code: string;
  tally_dt: Date;
  be_doc_no: string;
  master_ctn: number;
  loose_ctn: number;
  net_price: number;
  supp_ex_rate: number;
  local_charges_value: number;
  po_ex_rate: number;
  lc_po_value: number;
  po_curr_code: string;
  po_value: number;
  net_volume: number;
  hs_code: string;
  gross_wt: number;
  confirm_user: string;
  confirm_dt: Date;
  batch_no: string;
  created_at?: Date;
  created_by?: string;
  updated_at?: Date;
  updated_by?: string;
}
export interface QueryResultRow {
  error_code: number;
  error_message: string;
}